requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.0
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
Click==7.0
colorama==0.4.3
crayons==0.3.0
delegator.py==0.1.1
discord.py==1.3.3
docopt==0.6.2
idna==2.9
itsdangerous==1.1.0
lxml==4.5.0
MarkupSafe==1.1.1
multidict==4.7.5
pexpect==4.8.0
phlib==0.1.1
ptyprocess==0.6.0
requests==2.23.0
soupsieve==2.0
urllib3==1.25.9
websockets==8.1
Werkzeug==0.16.0
yarl==1.4.2
youtube-dl==2020.5.3
